Added the base version that we will expand upon
So far the game works between two players, takes score, and has a countdown timer before launching the ball again. We should be able to implement the ui with a begin test prompt before this, and set win limits easily. But anyway thats for future Alex and Anton to work on and those shmucks can figure that out.