Some minor updates and changes
changed a bunch of little things to get rid of player 2 functionality since its no longer a player. Added a method for the player 2 ai that has it follow the ball. In the main method the speed that it follows it is set to 10 atm. Its pretty good but will eventually lose to the player. Increasing the ball speed and its speed should give the player a harder time most likely. We can also look into shrinking the paddle size to make it harder for the player.